requirements.txt
amqp==5.2.0
anyio==4.1.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
black==23.11.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kombu==5.3.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==7.6.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
